Remove old quickstart, add dependencies
