+++ iam/api-client/pom.xml
@@ -55,6 +55,20 @@
       <version>v1-rev20200907-1.30.10</version><!-- v1 required here, v2 is different - DO NOT UPDATE to v2 -->
     </dependency>
 <!-- [END iam_java_quickstart_dependency] -->
+<!-- [START iamcredentials_java_dependency]-->
+    <dependency>
+      <groupId>com.google.apis</groupId>
+      <artifactId>google-api-services-iamcredentials</artifactId>
+      <version>v1-rev20200821-1.30.10</version>
+    </dependency>
+<!-- [END iamcredentials_java_dependency]-->
+<!-- [START troubleshooter_java_dependency]-->
+    <dependency>
+      <groupId>com.google.apis</groupId>
+      <artifactId>google-api-services-policytroubleshooter</artifactId>
+      <version>v1-rev20200801-1.30.10</version>
+    </dependency>
+<!-- [END troubleshooter_java_dependency]-->
     <dependency>
         <groupId>commons-cli</groupId>
         <artifactId>commons-cli</artifactId>
